pydbg.py: Use repr for actual exp value
Use `__repr__()` of expression value, as it's the intended representation for debugging.

For primitive types, this differentiates between `"True"` and `True`, and `"123"` and `123`, which can be crucial.

And for custom types, there are many using `__str__()` as part of their API for to-string conversion. This will remove the type details needed for debugging. `__repr__()` is usually created to preserve such details.