Clarify README
Link to info on dbg!

Not dbg! is not actually part of Rust 2018, but rather Rust-next, as you can see at
https://doc.rust-lang.org/edition-guide/rust-next/dbg-macro.html

Separate install from usage: move import, make example just work

Note that in Python you do have to import it, while in Rust
it is part of the standard prelude.